+++ src/app/ui/components/accordions/get-statiscts-accordion.component.ts
@@ -1,0 +1,28 @@
+import { Component, Input } from '@angular/core';
+import {
+  IonAccordion,
+  IonItem,
+  IonLabel,
+} from '@ionic/angular/standalone';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
+
+import { GetStatisticsComponent } from '../get-statistics/get-statistics.component';
+
+@Component({
+  selector: 'app-get-statistics-accordion',
+  templateUrl: './get-statistics-accordion.component.html',
+  standalone: true,
+  imports: [
+    IonAccordion,
+    IonItem,
+    IonLabel,
+    TranslateModule,
+    GetStatisticsComponent
+  ],
+})
+export class GetStatisticsAccordionComponent {
+  @Input() lang!: string;
+
+  constructor(public translate: TranslateService) {}
+
+}